requirements.txt
kubernetes
cryptography
pyyaml